Use code from k8s.io/utils repo
