+++ pkg/api/v1/service/util.go
@@ -21,15 +21,15 @@ import (
 	"strings"
 
 	"k8s.io/api/core/v1"
-	netsets "k8s.io/cloud-provider-openstack/pkg/util/net/sets"
+	"k8s.io/utils/net"
 )
 
 const (
 	defaultLoadBalancerSourceRanges = "0.0.0.0/0"
 )
 
-// IsAllowAll checks whether the netsets.IPNet allows traffic from 0.0.0.0/0
-func IsAllowAll(ipnets netsets.IPNet) bool {
+// IsAllowAll checks whether the net.IPNetSet allows traffic from 0.0.0.0/0
+func IsAllowAll(ipnets net.IPNetSet) bool {
 	for _, s := range ipnets.StringSlice() {
 		if s == "0.0.0.0/0" {
 			return true
@@ -41,13 +41,13 @@ func IsAllowAll(ipnets netsets.IPNet) bool {
 // GetLoadBalancerSourceRanges first try to parse and verify LoadBalancerSourceRanges field from a service.
 // If the field is not specified, turn to parse and verify the AnnotationLoadBalancerSourceRangesKey annotation from a service,
 // extracting the source ranges to allow, and if not present returns a default (allow-all) value.
-func GetLoadBalancerSourceRanges(service *v1.Service) (netsets.IPNet, error) {
-	var ipnets netsets.IPNet
+func GetLoadBalancerSourceRanges(service *v1.Service) (net.IPNetSet, error) {
+	var ipnets net.IPNetSet
 	var err error
 	// if SourceRange field is specified, ignore sourceRange annotation
 	if len(service.Spec.LoadBalancerSourceRanges) > 0 {
 		specs := service.Spec.LoadBalancerSourceRanges
-		ipnets, err = netsets.ParseIPNets(specs...)
+		ipnets, err = net.ParseIPNets(specs...)
 
 		if err != nil {
 			return nil, fmt.Errorf("service.Spec.LoadBalancerSourceRanges: %v is not valid. Expecting a list of IP ranges. For example, 10.0.0.0/24. Error msg: %v", specs, err)
@@ -59,7 +59,7 @@ func GetLoadBalancerSourceRanges(service *v1.Service) (netsets.IPNet, error) {
 			val = defaultLoadBalancerSourceRanges
 		}
 		specs := strings.Split(val, ",")
-		ipnets, err = netsets.ParseIPNets(specs...)
+		ipnets, err = net.ParseIPNets(specs...)
 		if err != nil {
 			return nil, fmt.Errorf("%s: %s is not valid. Expecting a comma-separated list of source IP ranges. For example, 10.0.0.0/24,192.168.2.0/24", v1.AnnotationLoadBalancerSourceRangesKey, val)
 		}
