+++ pkg/util/mount/mount.go
@@ -404,3 +404,13 @@ func getFileType(pathname string) (FileType, error) {
 
 	return pathType, fmt.Errorf("only recognise file, directory, socket, block device and character device")
 }
+
+// FileExists checks if specified file exists.
+func FileExists(filename string) (bool, error) {
+	if _, err := os.Stat(filename); os.IsNotExist(err) {
+		return false, nil
+	} else if err != nil {
+		return false, err
+	}
+	return true, nil
+}
